+++ contracts/capybara-finance-v2/test-utils/specific.ts
@@ -8,6 +8,6 @@ export interface Version {
 
 export const EXPECTED_VERSION: Version = {
   major: 2,
-  minor: 0,
+  minor: 2,
   patch: 0,
 };